fix: Prevent one from deleting their own account
Fixes #140.
